Newer versions of GCC (specifically, the one in MinGW64) don't like casting directly from a 64-bit pointer to a 32-bit integer, so we need to first cast to a 64-bit integer (remember that Windows is LLP64)
git-svn-id: svn://svn.code.sf.net/p/tigervnc/code/trunk@4722 3789f03b-4d11-0410-bbf8-ca57d06f2519

+++ win/winvnc/ControlPanel.cxx
@@ -147,11 +147,11 @@ void ControlPanel::SendCommand(DWORD command, int data)
     ListConnStatus.Copy(&ListSelConn);
     ListConnStatus.setAllStatus(data);
     ListConnStatus.setDisable(isItemChecked(IDC_DISABLE_CLIENTS));
-    copyData.cbData = (DWORD)&ListConnStatus;
+    copyData.cbData = (DWORD)((long long)&ListConnStatus);
   } else {
     ListConnStatus.Clear();
   }
-  copyData.cbData = (DWORD)&ListConnStatus;
+  copyData.cbData = (DWORD)((long long)&ListConnStatus);
   SendMessage(m_hSTIcon, WM_COPYDATA, 0, (LPARAM)&copyData);
 }
